Update deploy-to-lambda.yml workflow
This commit updates the deploy-to-lambda.yml workflow file. It adds environment variables for JavaScript source directory, package directory, package name, and Terraform directory. It also modifies the zip command to use the environment variables for source and destination paths. Finally, it updates the working directory for the Terraform apply step to use the environment variable.